# docs/design/personnel-mdm-openapi.yaml
openapi: 3.1.0
info:
  title: Lampang PAO - Central Personnel Master Data API
  version: 1.0.0
  summary: Single Source of Truth for personnel data of Lampang Provincial Administrative Organization (อบจ.ลำปาง)
  description: |
    API กลางสำหรับข้อมูลข้าราชการและพนักงาน อบจ.ลำปาง (Central Personnel MDM)

    **หลักการที่ผู้เรียกใช้ต้องทราบ**
    - ระบบปลายทางอ้างอิงบุคลากรด้วย `personId` (UUID) เท่านั้น เลขบัตรประชาชน (pid) จะไม่ปรากฏใน response ใด ๆ
      ยกเว้น `POST /persons/{personId}/pid` ซึ่งต้องมี scope `personnel:read:pid` และต้องระบุเหตุผลทุกครั้ง
    - การเข้าถึงข้อมูลควบคุมแบบ field-level ด้วย OAuth2 scope จาก Keycloak
      แต่ละกลุ่มฟิลด์ระบุ scope ที่ต้องใช้ไว้ใน `x-required-scope` ของ schema ฟิลด์ที่ไม่มีสิทธิ์จะถูกตัดออกจาก response (ไม่ใช่ null)
    - บุคคลสถานะ `INACTIVE` (ลาออก/โอนย้าย) เห็นได้เฉพาะ token ที่มี scope `personnel:read:inactive` ทั้งใน `GET /persons` และ endpoint รายบุคคล
      (`GET /persons/{id}`, `/photo`, `/employment`, `/change-log`, `/manage-profile`, `/history`, `POST /persons/{id}/pid`) ไม่มี scope -> 403 `insufficient-scope`
      (ไม่เขียน access log) บุคคลที่ไม่มีอยู่ยังได้ 404
    - ทุกคำขอที่คืนข้อมูลส่วนบุคคลถูกบันทึกใน access log (PDPA audit trail) พร้อม `purpose` ของระบบผู้เรียก
    - ข้อมูลระบุตัวตน (ชื่อ-สกุล วันเกิด ที่อยู่ตามทะเบียนบ้าน รูปถ่าย) มาจาก ThaID เท่านั้น แก้ไขผ่าน API ไม่ได้
    - เหตุการณ์เปลี่ยนแปลงข้อมูลส่งถึงระบบปลายทางสองทาง: webhook (push, มีลายเซ็น HMAC) และ `GET /events` (pull ด้วย cursor)
...

...
  contact:
...
servers:
...
security:
  - keycloak: []
...
paths:
...
  /persons:
    get:
      tags: [Persons]
      summary: Search personnel
      description: |
        ค้นหาบุคลากรด้วยชื่อ สังกัด ตำแหน่ง หรือสถานะ คืนเฉพาะฟิลด์ตาม scope ของผู้เรียก
        โดยค่าเริ่มต้นคืนเฉพาะ `status = ACTIVE` การดู record ที่ INACTIVE ต้องมี scope `personnel:read:inactive`
        ถ้าระบุ `status` ที่มี INACTIVE โดยไม่มี scope นี้ ตอบ 403 `insufficient-scope`
      operationId: searchPersons
      security:
        - keycloak: [personnel:read:basic]
      parameters:
        - name: q
          in: query
          description: |
            ค้นหาจากชื่อ-สกุล (ไทย/อังกฤษ) แบบ prefix match
            ไม่รองรับค้นหาด้วย employeeNo อีกต่อไป เพราะ employeeNo = เลขบัตรประชาชน (pid) เสมอ
            (อบจ.ลำปางไม่มีเลขประจำตัวข้าราชการแยกต่างหาก) การค้นหาด้วยค่านี้ผ่าน query string จะทำให้ pid
            หลุดไปอยู่ใน URL/access log ซึ่งขัดกฎข้อ 1 ของ CLAUDE.md
            ช่องว่างในชื่อ-สกุลส่งเป็น `%20` หรือ `+` ก็ได้ (`+` ถูก decode เป็นช่องว่าง) - พารามิเตอร์อื่นยังห้ามมี `+` ดิบ
          allowReserved: true
          schema: { type: string, minLength: 2, maxLength: 100 }
        - name: orgUnitId
          in: query
          description: กรองตามสังกัด (รวมหน่วยงานย่อยเมื่อ `includeChildUnits=true`)
          schema: { type: string, format: uuid }
        - name: includeChildUnits
          in: query
          schema: { type: boolean, default: false }
        - name: positionId
          in: query
          schema: { type: string, format: uuid }
        - name: personnelType
          in: query
          schema: { $ref: '#/components/schemas/PersonnelType' }
        - name: status
          in: query
          schema:
            type: array
            items: { $ref: '#/components/schemas/PersonStatus' }
            default: [ACTIVE]
          style: form
          explode: false
        - name: updatedSince
          in: query
          description: คืนเฉพาะ record ที่ `updatedAt` มากกว่าค่านี้ (ใช้ทำ incremental sync แบบง่าย)
          schema: { type: string, format: date-time }
        - $ref: '#/components/parameters/Fields'
        - $ref: '#/components/parameters/PidFormat'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: รายการบุคลากร (ฟิลด์ตาม scope)
          content:
            application/json:
              schema:
                type: object
                required: [data, page]
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Person' }
                  page: { $ref: '#/components/schemas/PageInfo' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [Provisioning]
      summary: Pre-provision a person (HR)
      description: |
        HR สร้าง record ล่วงหน้าสำหรับบุคลากรที่ยังไม่เคยเข้าสู่ระบบด้วย ThaID
        ระบบจะเก็บ pid ในรูป `pid_hash` (HMAC) และ `pid_enc` (เข้ารหัส) เท่านั้น และตั้งสถานะ `PENDING_CLAIM`
        เมื่อเจ้าตัวเข้าสู่ระบบด้วย ThaID ครั้งแรก record จะถูก claim อัตโนมัติและข้อมูลระบุตัวตนจะถูกเติมจาก ThaID

        - ถ้า `pid` ตรงกับ record ที่มีอยู่แล้ว จะคืน `409` พร้อม `personId` เดิม (ป้องกันข้อมูลซ้ำ)
        - `pid` ต้องผ่านการตรวจ checksum (mod 11) ก่อน

        **สิทธิ์ (เปลี่ยนใน PR-D1):** ต้องมี scope ของ operation นี้ **และ** realm role `hr_master_data_admin` (ตรวจที่ MDM API ทุก request
        เพราะ scope ใน token ไม่ใช่ตัวกั้นสิทธิ์) ไม่มี role -> 403 `insufficient-role` และต้องมี `reason` (บังคับ ห้ามมีเลข 13 หลัก:
        ว่าง -> 422 `reason-required`, มีเลข 13 หลัก -> 422 `reason-contains-pid`, เกิน 500 ตัวอักษร -> 422 `reason-too-long`) ซึ่งถูกเก็บใน
        `data_change_log.reason`
      operationId: provisionPerson
      security:
        - keycloak: [personnel:provision]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PersonProvisionRequest' }
      responses:
        '201':
          description: สร้าง record สถานะ PENDING_CLAIM แล้ว
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Person' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/Unprocessable' }
        '409':
          description: มี record ของ pid นี้อยู่แล้ว
          content:
            application/problem+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Problem'
                  - type: object
                    properties:
                      existingPersonId: { type: string, format: uuid }
  /persons/lookup:
    post:
      tags: [Persons]
      summary: Resolve pid to personId (restricted)
      description: |
        ใช้เฉพาะกรณีจำเป็น เช่น การเชื่อมข้อมูลครั้งแรกจากระบบเดิมที่ยังถือ pid อยู่
        คืนเฉพาะ `personId` และ `status` ไม่คืนข้อมูลอื่น ทุกครั้งถูกบันทึกใน access log พร้อม `justification`
        มี rate limit เข้มงวด (ค่าเริ่มต้น 30 ครั้ง/นาที/client) และคืน `404` ทั้งกรณีไม่พบและกรณีไม่มีสิทธิ์ดู record นั้น
        เพื่อป้องกันการใช้ endpoint นี้ตรวจสอบว่าใครเป็นบุคลากรหรือไม่ (enumeration)
      operationId: lookupPersonByPid
      security:
        - keycloak: [personnel:lookup:pid]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [pid, justification]
              properties:
                pid: { $ref: '#/components/schemas/Pid' }
                justification:
                  type: string
                  minLength: 10
                  maxLength: 500
                  description: เหตุผลการค้นหา บันทึกลง access log (ห้ามมีเลข 13 หลัก - ตอบ 422)
      responses:
        '200':
          description: พบ record
          content:
            application/json:
              schema:
                type: object
                required: [personId, status]
                properties:
                  personId: { type: string, format: uuid }
                  status: { $ref: '#/components/schemas/PersonStatus' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422':
          description: justification มีเลข 13 หลัก (ห้ามใส่เลขบัตรในเหตุผล)
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /persons/{personId}:
    parameters:
      - $ref: '#/components/parameters/PersonId'
    get:
      tags: [Persons]
      summary: Get a person by internal ID
      description: คืนข้อมูลบุคลากร เฉพาะกลุ่มฟิลด์ที่ผู้เรียกมี scope (ดู `x-required-scope` ใน schema `Person`)
      operationId: getPerson
      security:
        - keycloak: [personnel:read:basic]
      parameters:
        - $ref: '#/components/parameters/Fields'
        - $ref: '#/components/parameters/PidFormat'
        - name: If-None-Match
          in: header
          description: ETag จากครั้งก่อน (`"<version>"`) เพื่อรับ 304 เมื่อไม่มีการเปลี่ยนแปลง
          schema: { type: string }
      responses:
        '200':
          description: ข้อมูลบุคลากร
          headers:
            ETag:
              description: ค่า `version` ปัจจุบันของ record ในเครื่องหมายคำพูด
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Person' }
        '304': { description: ไม่มีการเปลี่ยนแปลงตั้งแต่ version ที่ระบุ }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /persons/{personId}/photo:
    parameters:
      - $ref: '#/components/parameters/PersonId'
    get:
      tags: [Persons]
      summary: Get the ThaID photo
      description: "คืนรูปถ่ายจากบัตรประชาชน (ThaID) เป็น binary ห้าม cache ที่ proxy (`Cache-Control: private, no-store`)"
      operationId: getPersonPhoto
      security:
        - keycloak: [personnel:read:photo]
      responses:
        '200':
          description: รูปถ่าย
          headers:
            X-Photo-SHA256:
              schema: { type: string }
              description: hash ของรูป ใช้ตรวจว่ารูปเปลี่ยนหรือไม่
          content:
            image/jpeg:
              schema: { type: string, format: binary }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /persons/{personId}/pid:
    parameters:
      - $ref: '#/components/parameters/PersonId'
    post:
      tags: [Persons]
      summary: Decrypt and return the national ID number (highly restricted)
      description: |
        คืนเลขบัตรประชาชนแบบถอดรหัสแล้ว สำหรับระบบที่มีฐานทางกฎหมายต้องใช้ pid เท่านั้น (เช่น ส่งข้อมูลให้กรมบัญชีกลาง กบข. หรือประกันสังคม)
        - ต้องมี scope `personnel:read:pid` ซึ่งอนุมัติเป็นราย client โดย DPO
        - ต้องส่ง `justification` ทุกครั้งใน JSON body (ไม่ส่งทาง query string เพื่อไม่ให้เหตุผลไปอยู่ใน URL/log ของ proxy) และถูกบันทึกใน access log แบบแยกตรวจสอบได้
        - เป็น `POST` ตั้งแต่ PR-B (เดิมเป็น `GET ...?justification=`; แถวเก่าใน access log ยังเป็น `GET`)
        - response ห้าม cache, ระบบผู้เรียกห้ามเก็บ pid ลง log ของตนเอง
      operationId: getPersonPid
      security:
        - keycloak: [personnel:read:pid]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [justification]
              additionalProperties: false
              properties:
                justification: { type: string, minLength: 10, maxLength: 500 }
      responses:
        '200':
          description: เลขบัตรประชาชน
          headers:
            Cache-Control:
              schema: { type: string, enum: ['private, no-store'] }
          content:
            application/json:
              schema:
                type: object
                required: [personId, pid]
                properties:
                  personId: { type: string, format: uuid }
                  pid: { $ref: '#/components/schemas/Pid' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422':
          description: justification มีเลข 13 หลัก (ห้ามใส่เลขบัตรในเหตุผล)
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
...
components:
  securitySchemes:
    keycloak:
      type: oauth2
      description: |
        Keycloak realm `lp-pao` เป็นผู้ออก token สำหรับเรียก MDM API (RS256, ตรวจสอบด้วย JWKS) โดยแยกจากเส้นทางล็อกอินของผู้ใช้
        ซึ่งยังคงเป็น thaid.lp-pao.go.th → check.lp-pao.go.th ตามเดิม
        - ระบบ ↔ ระบบ (รวม check.lp-pao.go.th และ backend ของทุกระบบปลายทาง) ใช้ **client credentials**
          และได้ scope ตาม client scope ที่ผู้ดูแล MDM/DPO อนุมัติเป็นราย client
        - ผู้ใช้ (self-service portal, HR/DPO console) ใช้ **authorization code + PKCE** กับ Keycloak
          ซึ่ง broker ไปยัง check.lp-pao.go.th ในฐานะ OIDC Identity Provider (ทางเลือก A ในเอกสารออกแบบ)
          และได้ scope ตาม realm role (`staff`, `hr_officer`, `dpo`)
        claim ที่ MDM ใช้: `sub`, `azp` (client id), `scope`, `person_id` (user context), `roles`
      flows:
        clientCredentials:
          tokenUrl: https://iam.lp-pao.go.th/realms/lp-pao/protocol/openid-connect/token
          scopes: &scopes
            personnel:read:basic: personId, ชื่อ-สกุล, ตำแหน่ง, สังกัด, ประเภทบุคลากร, สถานะ, version
            personnel:read:contact: เบอร์โทร, อีเมล, ที่อยู่ปัจจุบัน, ผู้ติดต่อฉุกเฉิน
            personnel:read:identity: วันเกิด, เพศ, ที่อยู่ตามทะเบียนบ้าน, วันออก/หมดอายุบัตร, สถานะ ThaID
            personnel:read:employment: วันบรรจุ, ระดับ, ประวัติการดำรงตำแหน่ง, เหตุพ้นจากราชการ
            personnel:read:photo: รูปถ่ายจาก ThaID
            personnel:read:inactive: เห็น record สถานะ INACTIVE (ลาออก/โอนย้าย)
            personnel:read:pid_masked: เห็นเลขบัตรประชาชนแบบปิด เหลือ 4 หลักท้าย (employeeNoMasked)
            personnel:read:pid: ถอดรหัสเลขบัตรประชาชน (อนุมัติโดย DPO เป็นราย client)
            personnel:lookup:pid: ค้นหา personId จาก pid (ใช้ตอน migrate/เชื่อมครั้งแรก)
            personnel:self: ข้อมูลของตนเอง และแก้ไขข้อมูลติดต่อ/ความยินยอม (user context เท่านั้น)
            personnel:write:employment: HR บันทึกตำแหน่ง/สังกัด/วันบรรจุ, ระงับ/คืนสถานะ, ขอ re-verify
            personnel:provision: HR เตรียม record ล่วงหน้าและจัดการ claim request
            personnel:import: นำเข้าข้อมูลชุดจากระบบ HR เดิม
            personnel:manage:reference: >-
              เพิ่ม/แก้ master data หน่วยงานและตำแหน่ง (T10) - ต้องมี realm role `hr_master_data_admin` ควบคู่เสมอ
              (API ตรวจทั้ง scope และ role เพราะ hr-console ผูก scope นี้ให้ทุกผู้ใช้ของ client)
            personnel:manage:person: >-
              HR จัดการข้อมูลบุคคลที่ไม่ใช่ของ ThaID (PR-D2: manage-profile, expected-identity, contact, emergency-contacts, history) - ต้องมี realm role
              `hr_master_data_admin` ควบคู่เสมอ (API ตรวจทั้ง scope และ role) ไม่ผูกกับฟิลด์ใดใน response ของ Person
            sync:thaid: check.lp-pao.go.th เท่านั้น (client `check-broker`)
            events:read: อ่านฟีดเหตุการณ์
            webhook:manage: จัดการ webhook subscription ของ client ตนเอง
            audit:read: อ่าน access log และ change log (DPO / ผู้ตรวจสอบ)
            audit:review: บันทึกผลรีวิวการเปิดเลขบัตร (เฉพาะ realm role dpo - MDM API ตรวจ role เพิ่มเสมอ ไม่ใช้ scope เป็นตัวกั้นเพียงอย่างเดียว)
        authorizationCode:
...
  parameters:
...
    PersonId:
      name: personId
      in: path
      required: true
      description: internal ID (UUID) ของบุคลากร
      schema: { type: string, format: uuid }
    Fields:
      name: fields
      in: query
      description: |
        เลือกกลุ่มฟิลด์ที่ต้องการ (comma-separated) เพื่อลดข้อมูลตาม data minimization
        ค่าที่ใช้ได้: basic, identity, contact, employment, emergencyContacts, verification
        ถ้าไม่ระบุ คืนทุกกลุ่มที่ scope อนุญาต
      schema: { type: string, examples: ['basic,employment'] }
    PidFormat:
      name: pidFormat
      in: query
      description: |
        `masked` = ไม่คืนเลขบัตรประชาชนเต็ม (`employeeNo`) แม้ token จะมี `personnel:read:pid` (data minimization ฝั่ง API:
        ผู้เรียกที่ต้องการแค่ `employeeNoMasked` ควรส่งค่านี้เสมอ) ไม่ส่ง = พฤติกรรมเดิมตาม scope
      schema: { type: string, enum: [masked] }
...
    Cursor:
      name: cursor
      in: query
      description: cursor จาก `page.nextCursor` ของหน้าก่อน
      schema: { type: string }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      description: UUID ที่ผู้เรียกสร้าง ระบบจะคืนผลเดิมถ้าเรียกซ้ำภายใน 24 ชั่วโมง
      schema: { type: string, format: uuid }
  responses:
    BadRequest:
      description: คำขอไม่ถูกต้อง
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Unauthorized:
      description: ไม่มี token หรือ token ไม่ถูกต้อง
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Forbidden:
      description: scope ไม่พอ หรือ client ถูกระงับ
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    NotFound:
      description: ไม่พบ (หรือไม่มีสิทธิ์เห็น record นั้น)
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
    Unprocessable:
      description: ข้อมูลถูกรูปแบบแต่ขัดกับ business rule (เช่น หน่วยงานที่อ้างถึงไม่มี/ไม่ active)
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    TooManyRequests:
      description: เกิน rate limit
      headers:
        Retry-After:
          schema: { type: integer }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
    Pid:
      type: string
      pattern: '^[0-9]{13}$'
      description: เลขประจำตัวประชาชน 13 หลัก (ต้องผ่าน checksum mod 11) ปรากฏเฉพาะใน request ของ endpoint ที่กำหนด
      x-data-classification: RESTRICTED
    PersonStatus:
      type: string
      enum: [PENDING_CLAIM, ACTIVE, INACTIVE]
    VerificationStatus:
      type: string
      enum: [UNVERIFIED, VERIFIED, STALE, EXPIRED]
      description: |
        UNVERIFIED = ยังไม่เคยเข้าสู่ระบบด้วย ThaID (record ที่ HR เตรียมไว้)
        VERIFIED = ข้อมูล ThaID ยืนยันภายในรอบที่กำหนด
        STALE = เกินรอบ ต้องยืนยันใหม่ (อยู่ใน grace period)
        EXPIRED = พ้น grace period แล้วยังไม่ยืนยัน
    PersonnelType:
      type: string
      description: |
        กฎเลขที่ตำแหน่ง (`positionId`) ตามประเภทบุคลากร - MDM API บังคับทุกเส้นทางที่เขียน employment
        (POST /persons, PUT /persons/{id}/employment, POST /persons/{id}/reactivate, POST /claim-requests/{id}/resolve
        action=PROVISION, POST /sync/hr/employment-batch) ผิดกฎ → `422`:
        - **ต้องมี** (`position-required`): `CIVIL_SERVANT`, `TEACHER`, `PERMANENT_EMPLOYEE`, `TRANSFERRED_HEALTH`
        - **ห้ามมี** (`position-not-allowed`): `CONTRACT_EMPLOYEE`, `GENERAL_EMPLOYEE`, `EXPERT_EMPLOYEE`,
          `OUTSOURCE_INDIVIDUAL`, `POLITICAL_APPOINTEE`
        - **ไม่บังคับ**: `OTHER` (และประเภทใหม่ที่ยังไม่ได้กำหนดกฎ)
        batch import รายงานเป็น error รายแถว `code = POSITION_REQUIRED | POSITION_NOT_ALLOWED`
      enum:
        - CIVIL_SERVANT            # ข้าราชการองค์การบริหารส่วนจังหวัด
        - TEACHER                  # ข้าราชการครูและบุคลากรทางการศึกษา
        - PERMANENT_EMPLOYEE       # ลูกจ้างประจำ
        - CONTRACT_EMPLOYEE        # พนักงานจ้างตามภารกิจ
        - GENERAL_EMPLOYEE         # พนักงานจ้างทั่วไป
        - EXPERT_EMPLOYEE          # พนักงานจ้างผู้เชี่ยวชาญพิเศษ
        - TRANSFERRED_HEALTH       # บุคลากรถ่ายโอน (รพ.สต.)
        - OUTSOURCE_INDIVIDUAL     # จ้างเหมาบริการ (รายบุคคล ประจำในหน่วยงาน ไม่มีเลขที่ตำแหน่ง)
        - OTHER                    # อื่นๆ (ไม่เข้ากลุ่มใดใน 8 ประเภทข้างต้น ไม่มีเลขที่ตำแหน่ง)
        - POLITICAL_APPOINTEE      # ผู้ดำรงตำแหน่งทางการเมือง (คณะผู้บริหาร - ไม่มีเลขที่ตำแหน่ง)
...
    Address:
      type: object
      properties:
        houseNo: { type: string }
        moo: { type: string }
        soi: { type: string }
        road: { type: string }
        subdistrict: { $ref: '#/components/schemas/AreaCode' }
        district: { $ref: '#/components/schemas/AreaCode' }
        province: { $ref: '#/components/schemas/AreaCode' }
        postcode: { type: string }
        fullText: { type: string, description: ที่อยู่แบบข้อความเต็ม }
    AreaCode:
      type: object
      properties:
        code: { type: string, description: รหัสตามกรมการปกครอง }
        nameTh: { type: string }
    Person:
      type: object
      description: |
        ข้อมูลบุคลากร แต่ละกลุ่มฟิลด์ปรากฏก็ต่อเมื่อผู้เรียกมี scope ตาม `x-required-scope`
        `basic` ปรากฏเสมอ (ต้องมี `personnel:read:basic`)
      required: [personId, status, version, updatedAt, basic]
      properties:
        personId: { type: string, format: uuid }
        status: { $ref: '#/components/schemas/PersonStatus' }
        version: { type: integer, description: เพิ่มขึ้นทุกครั้งที่ข้อมูลเปลี่ยน ใช้กับ ETag และเหตุการณ์ }
        updatedAt: { type: string, format: date-time }
        basic:
          type: object
          x-required-scope: personnel:read:basic
          properties:
            titleTh: { type: string, examples: [นาย] }
            firstNameTh: { type: string }
            lastNameTh: { type: string }
            titleEn: { type: string }
            firstNameEn: { type: string }
            lastNameEn: { type: string }
            employeeNo:
              type: string
              x-required-scope: personnel:read:pid
              description: |
                เลขประจำตัวข้าราชการ/พนักงาน = เลขบัตรประชาชนเสมอ (อบจ.ลำปางไม่มีเลขประจำตัวข้าราชการแยก
                ต่างหาก) จึงต้องมี scope personnel:read:pid เท่านั้น (เดียวกับ pid_hash/pid_enc)
            employeeNoMasked:
              type: string
              pattern: '^X-XXXX-XXXX\d-\d{2}-\d$'
              x-required-scope: personnel:read:pid_masked
              examples: ['X-XXXX-XXXX5-67-8']
              description: |
                เลขบัตรประชาชนที่ปิดด้วย API เหลือเฉพาะ 4 หลักท้าย (รูปแบบ X-XXXX-XXXX5-67-8) สำหรับให้เจ้าหน้าที่ยืนยันตัวตนบุคคล
                ต้องมี scope personnel:read:pid_masked - ไม่มี scope = ไม่มี key นี้ในผลลัพธ์ API ไม่ส่งเลขเต็มออกไปแล้วค่อยปิดที่ฝั่งผู้เรียก
            personnelType: { $ref: '#/components/schemas/PersonnelType' }
            positionTitle: { type: string }
            positionNo: { type: string }
            levelCode: { type: string, examples: [ชำนาญการ] }
            jobTitleText:
              type: string
              maxLength: 255
              description: ชื่อตำแหน่ง/ลักษณะงาน (ข้อความอิสระ) ของบุคลากรที่ไม่มีเลขที่ตำแหน่ง - ฟิลด์แยกจาก `positionTitle` ไม่ใช่ตัวแทน และไม่อยู่ใน tokenClaims
            orgUnit: { $ref: '#/components/schemas/OrgUnitRef' }
            emailWork: { type: string, format: email }
            photoUrl:
              type: string
              format: uri
              description: ลิงก์ `GET /persons/{personId}/photo` (ต้องมี scope photo จึงเรียกได้)
        verification:
          type: object
          x-required-scope: personnel:read:basic
          properties:
            verificationStatus: { $ref: '#/components/schemas/VerificationStatus' }
            thaidVerifiedAt: { type: [string, 'null'], format: date-time }
            claimedAt: { type: [string, 'null'], format: date-time }
        identity:
          type: object
          x-required-scope: personnel:read:identity
          x-source: THAID
          description: มาจาก ThaID เท่านั้น อ่านอย่างเดียว
          properties:
            middleNameTh: { type: [string, 'null'] }
            birthDate: { type: string, format: date }
            gender: { type: string, enum: [M, F, X] }
            registeredAddress: { $ref: '#/components/schemas/Address' }
            idCardIssueDate: { type: [string, 'null'], format: date }
            idCardExpireDate: { type: [string, 'null'], format: date }
            ial: { type: [string, 'null'], description: Identity Assurance Level จาก ThaID }
            syncedAt: { type: string, format: date-time }
        contact:
          $ref: '#/components/schemas/Contact'
        emergencyContacts:
          type: array
          x-required-scope: personnel:read:contact
          items: { $ref: '#/components/schemas/EmergencyContact' }
        employment:
          allOf:
            - $ref: '#/components/schemas/Employment'
          x-required-scope: personnel:read:employment
          description: employment ปัจจุบัน (ประวัติทั้งหมดดูที่ `GET /persons/{personId}/employment`)
    Contact:
      type: object
      x-required-scope: personnel:read:contact
      x-source: SELF
      properties:
        mobilePhone: { type: [string, 'null'] }
        phoneAlt: { type: [string, 'null'] }
        emailPersonal: { type: [string, 'null'], format: email }
        lineId: { type: [string, 'null'] }
        sameAsRegistered: { type: boolean }
        currentAddress: { $ref: '#/components/schemas/Address' }
        updatedAt: { type: string, format: date-time }
        updatedBy: { type: string, enum: [SELF, HR] }
...
    EmergencyContact:
      type: object
      required: [fullName, relationship, phone]
      properties:
        fullName: { type: string, minLength: 1, maxLength: 200 }
        relationship: { type: string, minLength: 1, maxLength: 50, examples: [คู่สมรส] }
        phone: { type: string, minLength: 1, maxLength: 20 }
        priority: { type: integer, minimum: 1, maximum: 3 }
    OrgUnitRef:
      type: object
      properties:
        orgUnitId: { type: string, format: uuid }
        code: { type: string }
        nameTh: { type: string }
        parentNameTh: { type: string, description: สำนัก/กอง ต้นสังกัด }
...
    Employment:
      type: object
      x-source: HR
      properties:
        employmentId: { type: string, format: uuid }
        employeeNo:
          type: string
          x-required-scope: personnel:read:pid
          description: เลขประจำตัวข้าราชการ/พนักงาน = เลขบัตรประชาชนเสมอ - ดู PersonDetail.basic.employeeNo
        personnelType: { $ref: '#/components/schemas/PersonnelType' }
        position: { $ref: '#/components/schemas/Position' }
        orgUnit: { $ref: '#/components/schemas/OrgUnitRef' }
        levelCode: { type: string }
        jobTitleText:
          type: string
          maxLength: 255
          x-required-scope: personnel:read:basic
          description: ชื่อตำแหน่ง/ลักษณะงาน (ข้อความอิสระ) - ดู `EmploymentUpsert.jobTitleText`
        appointedDate: { type: string, format: date, description: วันบรรจุ }
        effectiveFrom: { type: string, format: date }
        effectiveTo: { type: [string, 'null'], format: date }
        isCurrent: { type: boolean }
        employmentStatus:
          type: string
          enum: [ACTIVE, TRANSFERRED_OUT, RESIGNED, RETIRED, TERMINATED, DECEASED]
        separationDate: { type: [string, 'null'], format: date }
        emailWork: { type: [string, 'null'], format: email }
        hrSourceRef: { type: [string, 'null'], description: รหัสอ้างอิงในระบบ HR เดิม }
        updatedAt: { type: string, format: date-time }
    EmploymentUpsert:
      type: object
      required: [employeeNo, personnelType, orgUnitId, effectiveFrom]
      properties:
        employeeNo:
          type: string
          maxLength: 50
          description: เลขบัตรประชาชนของบุคคล (อบจ.ลำปางไม่มีเลขประจำตัวข้าราชการแยกต่างหาก)
        personnelType: { $ref: '#/components/schemas/PersonnelType' }
        positionId:
          type: string
          format: uuid
          description: |
            เลขที่ตำแหน่ง - ไม่ระบุได้สำหรับบุคลากรที่ไม่มีตำแหน่งทางการตามโครงสร้างอัตรากำลัง
            (พนักงานจ้าง/จ้างเหมาบริการรายบุคคล)
        orgUnitId: { type: string, format: uuid }
        levelCode: { type: string, maxLength: 50 }
        jobTitleText:
          type: string
          maxLength: 255
          description: |
            ชื่อตำแหน่ง/ลักษณะงาน (ข้อความอิสระ) สำหรับบุคลากรที่ไม่มีเลขที่ตำแหน่ง - เก็บใน `employment.job_title_text` แยกจาก
            `positionId` (ไม่แทน `basic.positionTitle` และไม่อยู่ใน token claims) ระบบตัด control/bidi characters, รวมบรรทัดเป็นบรรทัดเดียว
            และ trim ก่อนบันทึก (ข้อความว่างหลังตัด = ไม่มีข้อความ) กฎตามประเภทบุคลากร (แยกจากกฎเลขที่ตำแหน่ง):
            - ประเภทที่ต้องมีตำแหน่ง (`CIVIL_SERVANT`, `TEACHER`, `PERMANENT_EMPLOYEE`, `TRANSFERRED_HEALTH`) ที่ส่งฟิลด์นี้ -> 422 `job-title-not-allowed`
            - `OTHER` ส่งทั้ง `positionId` และ `jobTitleText` -> 422 `position-and-job-title-conflict` (เลือกอย่างใดอย่างหนึ่ง หรือไม่ใส่เลย)
            - ข้อความที่มีตัวเลข 13 หลักที่เข้าข่ายเลขบัตรประชาชน (ติดกันหรือมีขีด/เว้นวรรค) -> 422 `job-title-contains-pid`
            - ยาวเกิน 255 ตัวอักษรหลัง normalize -> 422 `job-title-too-long`
            (batch import รายงานเป็น error รายแถว: `JOB_TITLE_NOT_ALLOWED`, `POSITION_AND_JOB_TITLE_CONFLICT`, `JOB_TITLE_CONTAINS_PID`, `JOB_TITLE_TOO_LONG`)
        appointedDate: { type: string, format: date }
        effectiveFrom: { type: string, format: date }
        emailWork: { type: [string, 'null'], format: email, maxLength: 255 }
        referenceDocument: { type: string, maxLength: 200, description: เลขที่คำสั่ง }
        expectedVersion:
          type: integer
          description: version ของ person ที่ผู้เรียกเห็นล่าสุด (optimistic locking) ถ้าไม่ตรงคืน 409
...
    PersonProvisionRequest:
      type: object
      required: [pid, expectedFirstNameTh, expectedLastNameTh, employment, reason]
      properties:
        pid: { $ref: '#/components/schemas/Pid' }
        expectedBirthDate:
          type: string
          format: date
          description: วันเกิดที่ HR กรอก (รอยืนยันด้วย ThaID, ThaID เป็นหลักเสมอ) ช่วง 1900-01-01 ถึงวันนี้ (นอกช่วง -> 422)
        reason: { type: string, minLength: 1, maxLength: 500, description: เหตุผลที่สร้าง (บังคับ ห้ามมีเลข 13 หลัก) }
        expectedFirstNameTh: { type: string, minLength: 1, maxLength: 200 }
        expectedLastNameTh: { type: string, minLength: 1, maxLength: 200 }
        employment: { $ref: '#/components/schemas/EmploymentUpsert' }
        externalIds:
          type: array
          items:
            type: object
            required: [systemCode, value]
            properties:
              systemCode: { type: string }
              value: { type: string }
        notifyEmail:
          type: string
          format: email
          description: อีเมลสำหรับส่งคำเชิญให้เข้าสู่ระบบด้วย ThaID (ไม่เก็บเป็นข้อมูลติดต่อ)
...
    PageInfo:
      type: object
      properties:
        nextCursor: { type: [string, 'null'] }
        limit: { type: integer }
    Problem:
      type: object
      description: RFC 9457 Problem Details
      properties:
        type: { type: string, format: uri, examples: ['https://mdm.lp-pao.go.th/problems/insufficient-scope'] }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        requestId: { type: string }
        errors:
          type: array
          description: รายละเอียดรายฟิลด์สำหรับ 400
          items:
            type: object
            properties:
              field: { type: string }
              code: { type: string }
              message: { type: string }